// src/app/shared/model/cust-company-legal-doc-obj.model.ts
export class CustCompanyLegalDocObj {    
    CustCompanyLegalDocId: any;
    CustCompanyId: any;
    MrLegalDocTypeCode: any;
    MrLegalDocTypeDescr: string;
    DocNo: any;
    DocDt: any;
    DocExpiredDt: any;
    DocNotes: any;
    NotaryName: any;
    NotaryLocation: any;
    RowVersion: any;
    constructor(){this.CustCompanyLegalDocId = 0; this.RowVersion = "";}
}